pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amr-detection"
version = "1.0.0"
description = "Real-time AMR Detection System using YOLOv8 and Computer Vision"
readme = "README_NEW.md"
authors = [
    {name = "Rahul Ojha", email = "your.email@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "streamlit>=1.28.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "sphinx>=7.2.0",
    "notebook>=7.0.0",
]
gpu = [
    "torch>=2.0.0+cu118",
    "torchvision>=0.15.0+cu118",
]

[project.urls]
Homepage = "https://github.com/yourusername/amr-detection"
Documentation = "https://github.com/yourusername/amr-detection#usage"
Repository = "https://github.com/yourusername/amr-detection"
Issues = "https://github.com/yourusername/amr-detection/issues"

[project.scripts]
amr-detect = "src.streaming.app:main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --strict-markers"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
